add controllers and refactor code
