+++ models/contact.js
@@ -44,6 +44,7 @@ const contact = new Schema(
     owner: {
       type: Schema.Types.ObjectId,
       ref: 'user',
+      required: [true, "Owner Id is required"],
     }
   }
 );
